argparse handling multiple "--" in args improperly also 14364 - Argparse incorrectly handles '--'
This patch removes only one '--', the one that put a '-' in the
'arg_strings_pattern'.  It does this in 'consume_positionals' right
before calling 'take_action'.  As before it does not do this if
nargs is PARSER or REMAINDER.

It only removes the '--' that marked the end of options.

13922 is closed, but my patch refines it; it also fixes 14364
Maybe it should be billed as 14364 fix rather than 13922 reopen.