issue 11695 Improve argparse usage/help customization working idea that replaces 'parser.format_help()' with a 'custom_help'. This method takes a 'template', and fills it in with results from formatting usage, groups, description and epilog The default template replicates 'format_help' A test template is taken from the issue.
It includes a couple of contextmanger tests
Latest custom_help uses a decorator to handle repeated tasks
(that seems to be cleaner than creating a context)

'%(...)s' must be at start of a line, else multiline indentation gets messed up.
May need a more robust template tool

Is it better to keep this functionality in the parser, or move it to
the formatter (sub)class?