14174 - argparse.REMAINDER fails to parse remainder correctly
nargs REMAINDER is too greedy, at least by some user intuitions
If there is a positional, an optional, and a REMAINDER positional
the REMAINDER consume the optional

In this patch, the 1st REMAINDER string must be an 'A'.  In a sense
REMAINDER becomes a non-required PARSER.  It also includes a patch
from issue15112, that delays the consumption of positionals like '*'
to the end.

test_argparse.TestNargsRemainder is similar to the example raised in
the issue thread.  I added a test like the one that bothered the OP.

There is a slight chance that this will have backward compatibility problems.