'resolve' conflict handler alters the action in the parent parser
'resolve' tries to delete only conflicting option_strings, in the process
changing the Action itself.

sample3.py tries an alternative 'parent' handler.  If an action is shared
among 2 or more parsers (argument_groups) it is removed from the one where
it conflicts, without alteration.

_add_container_actions() adds a '.containers' attribute to record all of
the groups and parsers that the action is part of.